+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robot.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.kishi;
+package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.hardware.HardwareMap;
